TQS-25 Cleanned tests  files
