+++ backend/src/test/java/deti/tqs/backend/repositories/facility/RetrieveFacilityTests.java
@@ -1,4 +1,4 @@
-package deti.tqs.backend.repositories;
+package deti.tqs.backend.repositories.facility;
 
 import static org.junit.jupiter.api.Assertions.assertAll;
 import static org.assertj.core.api.Assertions.assertThat;
@@ -11,43 +11,27 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 
 import deti.tqs.backend.models.Facility;
+import deti.tqs.backend.repositories.FacilityRepository;
 
 @DataJpaTest
-class TestFacilityRepository {
-  
+public class RetrieveFacilityTests {
+
   private FacilityRepository facilityRepository;
 
   @Autowired
-  TestFacilityRepository(FacilityRepository facilityRepository) {
+  RetrieveFacilityTests(FacilityRepository facilityRepository) {
     this.facilityRepository = facilityRepository;
   }
 
-  @Test
-  @DisplayName("Test save a facility with success and find it by id")
-  void testSaveFacilityWithSuccess() {
-
-    Facility f = new Facility();
-
-    f.setName("Facility 1");
-    f.setCity("Aveiro");
-    f.setPhoneNumber("123456789");
-    f.setPostalCode("3810-193");
-    f.setStreetName("Rua de Aveiro");
-
-    facilityRepository.save(f);
-
-    Facility foundFacility = facilityRepository.findById(f.getId());
-
-    assertAll(
-      () -> assertThat(foundFacility).isNotNull(),
-      () -> assertThat(foundFacility.getName()).isEqualTo(f.getName()),
-      () -> assertThat(foundFacility.getCity()).isEqualTo(f.getCity()),
-      () -> assertThat(foundFacility.getPhoneNumber()).isEqualTo(f.getPhoneNumber()),
-      () -> assertThat(foundFacility.getPostalCode()).isEqualTo(f.getPostalCode()),
-      () -> assertThat(foundFacility.getStreetName()).isEqualTo(f.getStreetName())
-    );
-
-  }
+  /*
+   * NECESSARY TESTS
+   * 
+   *  1. Find all the facilities saved
+   *  2. Find a facility that doesn't exist
+   *  3. Find all locations
+   *  4. Find facilities by city
+   * 
+   */
 
   @Test
   @DisplayName("Test find all the facilities saved")
@@ -107,34 +91,6 @@ class TestFacilityRepository {
   }
 
   @Test
-  @DisplayName("Test save a facility with success and delete it")
-  void testSaveFacilityAndDeleteIt() {
-
-    Facility f = new Facility();
-    f.setName("Facility 1");
-    f.setCity("Aveiro");
-    f.setPhoneNumber("123456789");
-    f.setPostalCode("3810-193");
-    f.setStreetName("Rua de Aveiro");
-
-    facilityRepository.save(f);
-
-    Facility foundFacility = facilityRepository.findById(f.getId());
-
-    assertAll(
-      () -> assertThat(foundFacility).isNotNull(),
-      () -> assertThat(foundFacility.getName()).isEqualTo(f.getName())
-    );
-
-    facilityRepository.delete(f);
-
-    Facility deletedFacility = facilityRepository.findById(f.getId());
-
-    assertThat(deletedFacility).isNull();
-
-  }
-
-  @Test
   @DisplayName("Test find all locations")
   void testFindAllLocations() {
 
@@ -189,5 +145,5 @@ class TestFacilityRepository {
     );
 
   }
-
+  
 }
